+++ docs/pages/about.dox
@@ -1,0 +1,7 @@
+/**
+ * @page about About
+ * @section about-doxygen Doxygen documentation
+ * This page is auto generated using
+ * <a href="https://www.doxygen.nl/">Doxygen</a>, making use of some useful
+ * <a href="https://www.doxygen.nl/manual/commands.html">special commands</a>.
+ */
