+++ example/monitoring_tasks.cpp
@@ -1,0 +1,28 @@
+#include "BS_thread_pool.hpp"
+
+BS::synced_stream sync_out;
+BS::thread_pool pool(4);
+
+void sleep_half_second(const size_t i)
+{
+    std::this_thread::sleep_for(std::chrono::milliseconds(500));
+    sync_out.println("Task ", i, " done.");
+}
+
+void monitor_tasks()
+{
+    sync_out.println(pool.get_tasks_total(), " tasks total, ", pool.get_tasks_running(), " tasks running, ", pool.get_tasks_queued(), " tasks queued.");
+}
+
+int main()
+{
+    for (size_t i = 0; i < 12; ++i)
+        pool.push_task(sleep_half_second, i);
+    monitor_tasks();
+    std::this_thread::sleep_for(std::chrono::milliseconds(750));
+    monitor_tasks();
+    std::this_thread::sleep_for(std::chrono::milliseconds(500));
+    monitor_tasks();
+    std::this_thread::sleep_for(std::chrono::milliseconds(500));
+    monitor_tasks();
+}
