+++ example/pausing_workers.cpp
@@ -1,0 +1,30 @@
+#include "BS_thread_pool.hpp"
+
+BS::synced_stream sync_out;
+BS::thread_pool pool(4);
+
+void sleep_half_second(const size_t i)
+{
+    std::this_thread::sleep_for(std::chrono::milliseconds(500));
+    sync_out.println("Task ", i, " done.");
+}
+
+int main()
+{
+    for (size_t i = 0; i < 8; ++i)
+        pool.push_task(sleep_half_second, i);
+    sync_out.println("Submitted 8 tasks.");
+    std::this_thread::sleep_for(std::chrono::milliseconds(250));
+    pool.paused = true;
+    sync_out.println("Pool paused.");
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+    sync_out.println("Still paused...");
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+    for (size_t i = 8; i < 12; ++i)
+        pool.push_task(sleep_half_second, i);
+    sync_out.println("Submitted 4 more tasks.");
+    sync_out.println("Still paused...");
+    std::this_thread::sleep_for(std::chrono::milliseconds(1000));
+    pool.paused = false;
+    sync_out.println("Pool resumed.");
+}
